requirements.txt
psycopg2-binary
streamlit
pandas
plotly
matplotlib
os
hashlib